requirements.txt
Flask
mysql-connector-python
Werkzeug
openai
gTTS
opencc-python-reimplemented
